Move octokit initialization to db panel to send log messages to the user
